feat: :sparkles: 2. Gestión de Jugadores
feat: :sparkles: 2. Gestión de Jugadores

feat: :construction: Progress in validation and aggregation of a player in the implementation of case 2 of LigaBetPlay

The necessary validations of a player were carried out, such as his name and jersey number, so that when creating a new one, if it already exists, he will not be allowed to create it.

feat: :construction: Progress in the validation of updating and deleting a player in the implementation of LigaBetPlay case 2

In this preview we took into account to be able to update and delete a player from the database, that their ObjectId exists since this is the unique identifier to be able to make an update or deletion, if the player's id does not exist it will not allow updating or otherwise deleting any player

feat: :white_check_mark: Correct tests and validations to add, delete and update a player's data in the LigaBetPlay case 2 database

The respective tests of adding, deleting and updating data were carried out, taking into account their validations and everything went as expected by the administrator and the technical team

docs: :memo: Added documentation for use case 2